Porting over ActiveFedora adapter from Valkyrie
This copies code and tests from Valkyrie for the ActiveFedora adapter.

The only changes made were the addition of a predicate for description
so that our Collection resource will be created in Fedora. Additionally,
one test was marked pending because it passed individually but did not
pass in the suite.

This code (and commit) could be removed in the future if the
ActiveFedora adapter becomes part of the core Valkyrie gem.