Test for ActiveFedora and Fedora
Differentiates between testing ActiveFedora, which indexes in Solr and
writes to Fedora, and writing just to Fedora using Valkyrie's own
metadata adapter for Fedora.

In the latter case, a indexing persister is used, which enables us to
write to Fedora and index in Solr, but using only Valkyrie's adapters
and not using ActiveFedora at all.