Symlink custom Fedora config file
Valkyrie doesn't pass the Fedora username and password when making the
connection to the repository.

We can insert the credentials in the url used to connect, but that
requires a custom fedora.yml file. Since we're using the same server for
our other cho application, we can symlink the custom file and retain the
ability to switch back if we need.